converted FromWebMercator to use geom.Geometry; changed dependants of the tweaked basic.* functions as well
