+++ provider/gpkg/gpkg.go
@@ -86,20 +86,17 @@ func (p *Provider) TileFeatures(ctx context.Context, layer string, tile provider
 	// TODO(arolek): reimplement once the geom package has reprojection
 	// check if the SRID of the layer differs from that of the tile. tileSRID is assumed to always be WebMercator
 	if pLayer.srid != tileSRID {
-		minGeo, err := basic.FromWebMercator(pLayer.srid, basic.Point{tileBBox.MinX(), tileBBox.MinY()})
+		minGeo, err := basic.FromWebMercator(pLayer.srid, geom.Point{tileBBox.MinX(), tileBBox.MinY()})
 		if err != nil {
 			return fmt.Errorf("error converting point: %v ", err)
 		}
 
-		maxGeo, err := basic.FromWebMercator(pLayer.srid, basic.Point{tileBBox.MaxX(), tileBBox.MaxY()})
+		maxGeo, err := basic.FromWebMercator(pLayer.srid, geom.Point{tileBBox.MaxX(), tileBBox.MaxY()})
 		if err != nil {
 			return fmt.Errorf("error converting point: %v ", err)
 		}
 
-		tileBBox = &geom.Extent{
-			minGeo.AsPoint().X(), minGeo.AsPoint().Y(),
-			maxGeo.AsPoint().X(), maxGeo.AsPoint().Y(),
-		}
+		tileBBox = geom.NewExtent(minGeo.(geom.Point), maxGeo.(geom.Point))
 	}
 
 	var qtext string
